<Feat> 대부분의 공격에 원본 게임의 사운드 적용, FlyingAxe의 OrbitalCount 값을 실수 자료형으로 바꾸고 해당 값을 통해 원을 부분적으로 그릴 수 있도록 만듬
- Kratos 변수 정리

+++ Source/KingGodGeneralOfWar/Public/ANS_ThrowAxeInAttack.h
@@ -33,7 +33,7 @@ class KINGGODGENERALOFWAR_API UANS_ThrowAxeInAttack : public UAnimNotifyState
 	UPROPERTY(EditAnywhere, meta = (AllowPrivateAccess), Category = "Return")
 	bool bOrbital = false ;
 	UPROPERTY(EditAnywhere, meta = (AllowPrivateAccess), Category = "Return")
-	int OrbitalCount = 1;
+	float OrbitalCount = 1;
 	UPROPERTY(EditAnywhere, meta = (AllowPrivateAccess), Category = "Return")
 	bool bClockWise = true;
 	public:
